fix: reset requirement form on success
